% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PA_summary.R
\name{PA_summary}
\alias{PA_summary}
\title{Physical Activity Summary-Summations}
\usage{
PA_summary(final_dat)
}
\arguments{
\item{final_dat}{cleaned final data that is between record getup time and sleep time}
}
\value{
\code{hours_worn_total} Total worn hours

\code{sed_hour}  Total sedantary hours

\code{stand_hour} Total stand hours

\code{step_hour} Total step hours

\code{num_changes_from_sed_to_non_sed} Numbers of changing from sedantary to non sadantary status

\code{step_count_total} Total Step count

\code{num_hour_over_3_METs} Number of hours that Metabolic Equivalent of Task (MET) is over 3

\code{MET_hours}  Total METs

\code{table} A table that shows all above results
}
\description{
Summarize activity measures
}
\details{
All numbers are calculated in the given time period (day, hour, etc_)_ Total sedentary/standing/stepping hours are obtained from the summation of the duration times for sedentary/standing/stepping activities in the given time period_
}
\examples{
 data(sampledata);PA_summary(sampledata)
}

