% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_bed_time.r
\docType{data}
\name{sample_bed_time}
\alias{sample_bed_time}
\title{Simulated Sample Sleep and Wake up Time Log}
\format{A data frame with 7 rows and 6 variables}
\usage{
data(sample_bed_time)
}
\description{
This log matches event data with the same 7 consecutive days. The variables are as follows:
\itemize{
  \item \code{id} The id of participant
  \item \code{Visit} The visit number
  \item \code{date.out} Wake up date
  \item \code{time.out} Wake up time
  \item \code{date.in} Sleep date
  \item \code{time.in} Sleep time
}
}
\keyword{datasets}

