% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_event.r
\docType{data}
\name{sample_event}
\alias{sample_event}
\title{Simulated Sample Accelerometer Data}
\format{A data frame with 21687 rows, and 6 variables}
\usage{
data(sample_event)
}
\description{
Seven consecutive days data is provided.
The variables are as follows:
\itemize{
  \item \code{Time} Time of the recored physical activity from simmulated ActivPAL data
  \item \code{Data Count} Data Count (samples)
  \item \code{Time interval} (in second)
  \item \code{Activity Code} 0-sedantary, 1-standing, 2-stepping
  \item \code{Cumulative Step Count}
  \item \code{Activity Score} (MET.h)
}
}
\keyword{datasets}

