% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.r
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{Simulated sample accelerometer data}
\format{A data frame with 128 rows and 6 variables}
\usage{
data(sampledata)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\details{
\itemize{
  \item Time Time of the recored physical activity from simmulated ActivPAL data
  \item Data Count Data Count (samples)
  \item Time interval (in second)
  \item Activity Code 0-sedantary, 1-standing, 2-stepping
  \item Cumulative Step Count
  \item Activity Score (MET.h)
}
}
\keyword{datasets}

