% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_bed_time.r
\docType{data}
\name{sample_bed_time}
\alias{sample_bed_time}
\title{Simulated sample Sleep and wake up time log}
\format{A data frame with 7 rows and 6 variables}
\usage{
data(sample_bed_time)
}
\description{
The variables are as follows:
\itemize{
  \item id The id of participant
  \item Visit The visit number
  \item date.out Wake up date
  \item time.out Wake up time
  \item date.in Sleep date
  \item time.in Sleep time
}
}
\keyword{datasets}

