% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_takeon_log.r
\docType{data}
\name{sample_takeon_log}
\alias{sample_takeon_log}
\title{Simulated sample take on and off log}
\format{A data frame with 7 rows and 6 variables}
\usage{
data(sample_takeon_log)
}
\description{
The variables are as follows:
\itemize{
  \item id The id of participant
  \item Visit The visit number
  \item date.on Take on date
  \item time.on Take on time
  \item date.off Take off date
  \item time.off Take off time
}
}
\keyword{datasets}

